Moved models into dedicated folder
