+++ frontend/scripts/game/models/Paddle.js
@@ -1,4 +1,4 @@
-import {CANVAS_HEIGHT, CANVAS_WIDTH} from "../utils/constants.js";
+import {CANVAS_HEIGHT} from "../../utils/constants.js";
 
 export default class Paddle {
     constructor(x, y, width, height, speed) {
